+++ src/frontend/js/components/DashboardItem/DashboardItemEnrollment.stories.tsx
@@ -1,0 +1,31 @@
+import { ComponentMeta, ComponentStory } from '@storybook/react';
+import { DashboardItemEnrollment } from './DashboardItemEnrollment';
+import { enrollment } from './mock.stories';
+
+export default {
+  title: 'Components/Dashboard/Enrollment',
+  component: DashboardItemEnrollment,
+} as ComponentMeta<typeof DashboardItemEnrollment>;
+
+const Template: ComponentStory<typeof DashboardItemEnrollment> = (args) => (
+  <DashboardItemEnrollment {...args} />
+);
+
+export const Opened = Template.bind({});
+Opened.args = {
+  enrollment: {
+    ...enrollment,
+    course_run: { ...enrollment.course_run, end: '2029-10-01T01:23:37+00:00' },
+  },
+};
+
+export const Closed = Template.bind({});
+Closed.args = {
+  enrollment: {
+    ...enrollment,
+    course_run: { ...enrollment.course_run, end: '2021-10-01T01:23:37+00:00' },
+  },
+};
+
+export const NotActive = Template.bind({});
+NotActive.args = { enrollment: { ...enrollment, is_active: false } };